# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  id: number;
  userId: string;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageKey?: string | null;
  createdAt: string;
}