# lib/api-zod/src/generated/types/paymentIntentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentIntentInput {
  amount: number;
  currency: string;
}