# lib/api-zod/src/generated/types/questionnaire.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionnaireStatus } from './questionnaireStatus';

export interface Questionnaire {
  id: number;
  userId: string;
  currentStep: number;
  status: QuestionnaireStatus;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  birthTime?: string | null;
  /** @nullable */
  birthCity?: string | null;
  /** @nullable */
  question1?: string | null;
  /** @nullable */
  question2?: string | null;
  /** @nullable */
  question3?: string | null;
  /** @nullable */
  question4?: string | null;
  /** @nullable */
  question5?: string | null;
  createdAt: string;
  updatedAt?: string;
}