# lib/api-zod/src/generated/types/questionnaireStepInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionnaireStepInput {
  step: number;
  birthDate?: string;
  birthTime?: string;
  birthCity?: string;
  question1?: string;
  question2?: string;
  question3?: string;
  question4?: string;
  question5?: string;
}