# lib/api-zod/src/generated/types/questionnaireWithTeaser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Questionnaire } from './questionnaire';
import type { TeaserSummary } from './teaserSummary';

export interface QuestionnaireWithTeaser {
  questionnaire: Questionnaire;
  teaser: TeaserSummary;
}