# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Astrology Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  name?: string;
  avatarUrl?: string;
}